Isolate frame from factory (#2284)
