+++ src/R/Components/Impl/InteractiveWorkflow/Implementation/RInteractiveWindowVisualComponent.cs
@@ -3,15 +3,20 @@
 
 using System;
 using System.Windows;
+using Microsoft.Common.Core;
 using Microsoft.R.Components.Controller;
 using Microsoft.R.Components.Services;
 using Microsoft.R.Components.View;
+using Microsoft.R.Host.Client;
 using Microsoft.VisualStudio.InteractiveWindow;
 using Microsoft.VisualStudio.Text;
 using Microsoft.VisualStudio.Text.Editor;
+using static System.FormattableString;
 
 namespace Microsoft.R.Components.InteractiveWorkflow.Implementation {
     public class RInteractiveWindowVisualComponent : IInteractiveWindowVisualComponent {
+        private readonly IRSessionProvider _sessionProvider;
+
         public ICommandTarget Controller { get; }
         public FrameworkElement Control { get; }
         public IInteractiveWindow InteractiveWindow { get; }
@@ -22,18 +27,41 @@ namespace Microsoft.R.Components.InteractiveWorkflow.Implementation {
 
         public IVisualComponentContainer<IVisualComponent> Container { get; }
 
-        public RInteractiveWindowVisualComponent(IInteractiveWindow interactiveWindow, IVisualComponentContainer<IInteractiveWindowVisualComponent> container) {
+        public RInteractiveWindowVisualComponent(IInteractiveWindow interactiveWindow, IVisualComponentContainer<IInteractiveWindowVisualComponent> container, IRSessionProvider sessionProvider) {
             InteractiveWindow = interactiveWindow;
             Container = container;
+
+            _sessionProvider = sessionProvider;
+            sessionProvider.BrokerChanged += OnBrokerChanged;
 
             var textView = interactiveWindow.TextView;
             Controller = ServiceManagerBase.GetService<ICommandTarget>(textView);
             Control = textView.VisualElement;
             interactiveWindow.Properties.AddProperty(typeof(IInteractiveWindowVisualComponent), this);
+
+            UpdateWindowTitle();
         }
 
         public void Dispose() {
             InteractiveWindow.Properties.RemoveProperty(typeof (IInteractiveWindowVisualComponent));
+            _sessionProvider.BrokerChanged -= OnBrokerChanged;
+        }
+
+        private void OnBrokerChanged(object sender, EventArgs e) {
+            UpdateWindowTitle();
+        }
+
+        private void UpdateWindowTitle() {
+            var broker = _sessionProvider.Broker;
+            string title;
+            if (broker != null) {
+                title = broker.IsRemote
+                    ? Invariant($"{Resources.ReplWindowName} - {broker.Name} ({broker.Uri.ToString().TrimTrailingSlash()})")
+                    : Invariant($"{Resources.ReplWindowName} - {broker.Name}");
+            } else {
+                title = Resources.Disconnected;
+            }
+            Container.CaptionText = title;
         }
     }
 }
